feat: add createBankAccount API and related types
- Introduced `createBankAccount` API in `createBankAccount.ts` to facilitate bank account creation.
- Added corresponding types `CreateBankAccountPayload` and `CreateBankAccountResponse`.
- Updated `apis.ts` and `index.ts` to include the new API method and types.
- Implemented a utility function `normalizeHolderName` in `utils.ts` for holder name normalization.
- Enhanced `BankAccount` model to accommodate optional fields for better flexibility.
