+++ src/index.ts
@@ -21,6 +21,7 @@ export type { ChangeGroupAvatarResponse } from "./apis/changeGroupAvatar.js";
 export type { ChangeGroupNameResponse } from "./apis/changeGroupName.js";
 export type { ChangeGroupOwnerResponse } from "./apis/changeGroupOwner.js";
 export type { CreateAutoReplyPayload, CreateAutoReplyResponse } from "./apis/createAutoReply.js";
+export type { CreateBankAccountPayload, CreateBankAccountResponse } from "./apis/createBankAccount.js";
 export type { CreateCatalogResponse } from "./apis/createCatalog.js";
 export type { CreateGroupOptions, CreateGroupResponse } from "./apis/createGroup.js";
 export type { CreateNoteOptions, CreateNoteResponse } from "./apis/createNote.js";
